# specs/002-pdp-product/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: PDP P1 平台 API
  version: 1.1.0
  description: |
    PDP 企业级项目交付管理平台 P1 的稳定同步接口边界。
    本契约与 coverage.md 共同覆盖全部 P1 稳定操作；领域包生成页面使用
    domain-package.schema.json 定义的元数据，不为客户复制私有接口。
...
servers:
  - url: /api/v1
security:
  - oidc: []
...
paths:
...
  /workflow-definitions/validate:
    post:
      tags: [平台工作流]
      summary: 校验 BPMN 流程定义
      operationId: validateWorkflowDefinition
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [processDefinitionKey, businessVersion, bpmnXml]
              properties:
                processDefinitionKey:
                  type: string
                  pattern: '^[a-z][a-z0-9.-]{2,99}$'
                businessVersion:
                  type: string
                  pattern: '^[0-9]+\.[0-9]+\.[0-9]+$'
                domainPackageVersionId:
                  oneOf:
                    - $ref: '#/components/schemas/Uuid'
                    - type: 'null'
                bpmnXml:
                  type: string
                  minLength: 50
      responses:
        '200':
          description: 校验结果
          content:
            application/json:
              schema:
                type: object
                required: [valid, contentHash, findings]
                properties:
                  valid:
                    type: boolean
                  contentHash:
                    type: string
                  findings:
                    type: array
                    items:
                      type: object
                      required: [severity, code, message]
                      properties:
                        severity:
                          type: string
                          enum: [ERROR, WARNING]
                        code:
                          type: string
                        message:
                          type: string
  /workflow-definitions/deploy:
    post:
      tags: [平台工作流]
      summary: 部署已批准的 BPMN 流程定义
      operationId: deployWorkflowDefinition
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [processDefinitionKey, businessVersion, contentHash, bpmnResource]
              properties:
                processDefinitionKey:
                  type: string
                businessVersion:
                  type: string
                contentHash:
                  type: string
                bpmnResource:
                  type: string
                domainPackageVersionId:
                  oneOf:
                    - $ref: '#/components/schemas/Uuid'
                    - type: 'null'
      responses:
        '201':
          description: 已部署
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowDefinitionSummary'
        '409':
          $ref: '#/components/responses/Problem'
  /workflow-instances/{workflowInstanceId}:
    get:
      tags: [平台工作流]
      summary: 查询平台流程实例诊断摘要
      operationId: getWorkflowInstance
      parameters:
        - name: workflowInstanceId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/Uuid'
      responses:
        '200':
          description: 流程实例摘要
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowInstanceSummary'
        '404':
          $ref: '#/components/responses/Problem'
  /workflow-instances/{workflowInstanceId}/actions:
    post:
      tags: [平台工作流]
      summary: 执行受控流程管理动作
      operationId: applyWorkflowAdministrationAction
      parameters:
        - name: workflowInstanceId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/Uuid'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkflowAdminActionCommand'
      responses:
        '202':
          description: 管理动作已接受
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobAccepted'
        '409':
          $ref: '#/components/responses/Problem'
...
components:
  securitySchemes:
    oidc:
      type: openIdConnect
      openIdConnectUrl: https://idp.example.com/.well-known/openid-configuration
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 16
        maxLength: 128
...
  responses:
    Problem:
      description: 标准问题响应
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
  schemas:
    Uuid:
      type: string
      format: uuid
...
    Problem:
      type: object
      required: [type, title, status, traceId]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
        code:
          type: string
        traceId:
          type: string
        violations:
          type: array
          items:
            type: object
            required: [field, code]
            properties:
              field:
                type: string
              code:
                type: string
              message:
                type: string
...
    WorkflowDefinitionSummary:
      type: object
      required: [id, processDefinitionKey, businessVersion, contentHash, status, deployedAt]
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        processDefinitionKey:
          type: string
        businessVersion:
          type: string
        contentHash:
          type: string
        domainPackageVersionId:
          oneOf:
            - $ref: '#/components/schemas/Uuid'
            - type: 'null'
        status:
          type: string
          enum: [VALIDATED, DEPLOYED, DEPRECATED, RETIRED]
        deployedAt:
          type: string
          format: date-time
    WorkflowInstanceSummary:
      type: object
      required: [id, definitionId, businessObjectRef, state, revision, incidentCount]
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        definitionId:
          $ref: '#/components/schemas/Uuid'
        businessObjectRef:
          type: object
          additionalProperties: true
        state:
          type: string
          enum: [STARTING, ACTIVE, SUSPENDED, COMPLETED, TERMINATED, INCIDENT]
        currentActivityKeys:
          type: array
          items:
            type: string
        incidentCount:
          type: integer
          minimum: 0
        revision:
          type: integer
          minimum: 0
        incidents:
          type: array
          description: 用于平台诊断的 incident 与 dead-letter 摘要；业务接口不得直接暴露引擎表。
          items:
            $ref: '#/components/schemas/WorkflowIncidentSummary'
    WorkflowIncidentSummary:
      type: object
      required: [id, code, message, status, attempts, occurredAt]
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        code:
          type: string
        message:
          type: string
        status:
          type: string
          enum: [OPEN, RETRYING, RESOLVED, DEAD_LETTER]
        attempts:
          type: integer
          minimum: 0
        occurredAt:
          type: string
          format: date-time
        resolvedAt:
          type: [string, 'null']
          format: date-time
    WorkflowAdminActionCommand:
      type: object
      required: [action, reason, expectedRevision]
      properties:
        action:
          type: string
          enum: [PAUSE, RESUME, RETRY, MIGRATE, TERMINATE, MANUAL_COMPENSATE]
        reason:
          type: string
          minLength: 5
          maxLength: 2000
        expectedRevision:
          type: integer
          minimum: 0
        targetDefinitionId:
          oneOf:
            - $ref: '#/components/schemas/Uuid'
            - type: 'null'
        impactPreviewId:
          oneOf:
            - $ref: '#/components/schemas/Uuid'
            - type: 'null'
    JobAccepted:
      type: object
      required: [jobId, status]
      properties:
        jobId:
          $ref: '#/components/schemas/Uuid'
        status:
          const: QUEUED
...